Propagate aliased names forward when aliasing a block
If the current block becomes an alias for another, but it too is an alias for an earlier block, then
the earlier block should be forwarded to the latter.
e.g:
LabelA:
LabelB:
LabelC:

LabelA aliases LabelB and LabelB aliases LabelC. (LabelA -> LabelB -> LabelC)
We can also say that LabelA aliases LabelC.
Previously LabelC was not aware it had LabelA as an alias, therefore it was not safe
to eliminate LabelA. This change allows LabelC to remove LabelA provided LabelA has no references.
Additonally, as each entry block was forwarded through the alias, it would increment
the refcount for the forwarder, so A -> C via B incremented B. enterByName was added
to correctly increment A instead.

Fixes #9